requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
filelock==3.14.0
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==11.2.1
platformdirs==4.2.1
PyJWT==2.8.0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
virtualenv==20.26.2
whitenoise==6.9.0
